Style the cta section, refactor icons and buttons
For icons, I refactored the `icon-chevron` class to `icon-stroke` so it
can be used for all icons that use stroke instead of fill.

I also adjusted the alignment of the svg in the intro section, since the
previous svg we used is a fill type and the new svg was a bit off center.

For the buttons, I made the css more robust for light and dark
backgrounds. By default, buttons will have a border color of
`currentColor`. I used a custom data attribute `[data-bg="dark"]` to
indicate buttons with dark-colored backgrounds, so they can be styled
appropriately on hover.