Use `typeset` for improved typography
This replaces `eleventy-plugin-typography` which has a bug that leaves
double quotes as straight quotes.

BTW there's another community plugin `eleventy-plugin-typeset` which is
basically a wrapper around `typeset`, so I decided to use the source.

I disabled two features since they break the layout in certain sections:
- hyphenate
- hangingPunctuation

`typeset` works as a transform instead of a filter, so I updated the
affected templates and markdown files to use the new approach.

Lastly, I used a non-breaking space in the intro for the phrase "web
developer" to make the layout more consistent across different screen
sizes.